restructuring and cleanup of codebase
