+++ src/Logic Functions/tripProcessor.js
@@ -1,4 +1,3 @@
-//LOGIC FUNCTIONS 
 const filterTripsByTraveler = (trips = [], travelerId) => {
   return trips.filter(trip => trip.userID === travelerId);
 }
@@ -21,7 +20,6 @@ const categorizeTrips = (trips) => {
   return { pastTrips, upcomingTrips, pendingTrips };
 }
 
-
 const calculateTotalAmountSpent = (pastTrips = [], destinations = [], currentYear) => {
   const agentFee = 0.10;
   let totalAmountSpent = 0;
@@ -41,11 +39,8 @@ const calculateTotalAmountSpent = (pastTrips = [], destinations = [], currentYea
           }
       }
   });
-
   return totalAmountSpent;
 };
-
-
 
 const getTripDetailsForTraveler = (travelerId, trips = [], destinations = [], currentYear = 2020) => {
   try {
