+++ src/css/styles.css
@@ -331,7 +331,7 @@ footer {
 }
 
 .calculate-cost-button, .submit-request-button{ 
-  background-color: #E35D6A; 
+  background-color: #004080;
   font-size: 17px;
   border-radius: 5px;
   color: white;
